+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-## 2.10
+## develop
+
+* Foundation food improvements:
+  * Bias foundation food matching to prefer "raw" FDC ingredients, but only if the ingredient name does not include any verbs that indicate the ingredient is not raw (e.g. "cooked").
+  * Normalise spelling of tokens in ingredient names to align with spelling used in FDC ingredient descriptions.
+
+## 2.1.1
+
+* Pin Pint version to 0.24.4, as future versions intend to drop support for Python 3.10.
+
+## 2.1.0
 
 > [!WARNING]
 >